added jwt authentication for using api's

+++ controllers/api/v1/posts_api.js
@@ -21,13 +21,19 @@ module.exports.index=async function(req,res){
 module.exports.destroy=async function(req,res){
     try{
         let post=await Post.findById(req.params.id);
-    
+        
+        if(post.user == req.user.id){
             post.remove();
             await Comment.deleteMany({post:req.params.id});
 
             return res.json(200,{
                 message:"Post deleted with associated comments"
             });
+        }else{
+            return res.json(401,{
+                message:"you cannot delete this post!"
+            });
+        }
     }
     catch(err){
         console.log("error occurred:",err);
